Switched from .txt to .edn.
Wins/losses/draws are saved to an .edn and they persist. The player’s
total tally is displayed when the user selects see scores. Scores from
various games and match-ups are combined to great a total tally, that
is the entire history of the player’s game win/loss/draw history.