Added the ability to save data to either a EDN or JSON file.
Also designed repository structure so that it is easy to add another
method of saving data. Score recording is now an interface for the
repository that gains the information of player names, total tally and
is the front for recording it. Added tests for the EDN and JSON only in
the scoring repository, as different contexts.